Update script to convert data to demo-empirical-data schema
This mainly involved changin some class names, namespaces, and urls.
Also, it was decided not to generate display_name (i.e. preflabel)
values for all records, so that these can rather be defined by the
shacl-vue config option for templating display names.

Lastly, the readme was updated to include an improved description
for how to use the script to generate and POST metadata to a dumpthings
backend, and how to delete existing data beforehand and restart the
service afterwards.